src/nathandloria/requirements.txt
asgiref==3.3.0
dj-database-url==0.5.0
Django==3.1.3
django-heroku==0.3.1
django-solo==1.1.3
django-widget-tweaks==1.4.8
gunicorn==20.0.4
psycopg2==2.8.6
python-dotenv==0.15.0
pytz==2020.4
sqlparse==0.4.1
whitenoise==5.2.0
